Moved some tests from skill_model_test to wa_functions_test.
