+++ skillmodels/tests/estimation/skill_model_test.py
@@ -11,8 +11,6 @@ from pandas.util.testing import assert_frame_equal
 from pandas.util.testing import assert_series_equal
 
 from skillmodels import SkillModel as smo
-from skillmodels.estimation.wa_functions import all_variables_for_iv_equations, variable_permutations_for_iv_equations, \
-    number_of_iv_parameters, extended_meas_coeffs, residual_measurements, model_coeffs_from_iv_coeffs_args_dict
 
 
 class TestGeneralParamsSlice:
@@ -904,177 +902,6 @@ class TestSigmaWeightsAndScalingFactor:
 class TestLikelihoodArgumentsDict:
     def setup(self):
         pass
-
-
-class TestAllVariablesForIVEquations:
-    def setup(self):
-        self.measurements = {
-            'f1': [['y01', 'y02'], ['y11', 'y12'], []],
-            'f2': [['y04', 'y05'], ['y14', 'y15'], []],
-            'f3': [['y07', 'y08'], [], []]}
-
-        self.factors = ['f1', 'f2', 'f3']
-        self.included_factors = [['f1', 'f3'], ['f2', 'f3'], []]
-        self.transition_names = ['blubb', 'blubb', 'constant']
-        self.periods = [0, 1, 2]
-
-    def test_all_variables_for_iv_equations_constant_factor(self):
-        calc_meas_list = all_variables_for_iv_equations(self.factors, self.included_factors, self.transition_names,
-                                                        self.measurements, 1, 'f1')
-        expected_meas_list = [
-            ['y11_resid', 'y12_resid'],
-            ['y07_copied_resid', 'y08_copied_resid']]
-        assert_equal(calc_meas_list, expected_meas_list)
-
-    def test_all_variables_for_iv_equations_non_constant(self):
-        calc_meas_list = all_variables_for_iv_equations(self.factors, self.included_factors, self.transition_names,
-                                                        self.measurements, 1, 'f2')
-        expected_meas_list = [
-            ['y14_resid', 'y15_resid'], ['y07_copied_resid', 'y08_copied_resid']]
-        assert_equal(calc_meas_list, expected_meas_list)
-
-    @patch('skillmodels.estimation.wa_functions.all_variables_for_iv_equations')
-    def test_indepvar_permutations(self, mock_allvars):
-        ret_val = [['y1', 'y2'], ['y3', 'y4']]
-        self.anchored_factors = []
-        mock_allvars.return_value = ret_val
-
-        expected_xs = [
-            ['y1', 'y3'], ['y1', 'y4'], ['y2', 'y3'], ['y2', 'y4']]
-        calc_xs = variable_permutations_for_iv_equations(self.factors, self.included_factors, self.transition_names,
-                                                         self.measurements, self.anchored_factors, 1, 1)[0]
-        assert_equal(calc_xs, expected_xs)
-
-    @patch('skillmodels.estimation.wa_functions.all_variables_for_iv_equations')
-    def test_instrument_permutations(self, mock_allvars):
-        self.anchored_factors = []
-        ret_val = [['y1_resid', 'y2_resid'], ['y3_resid', 'y4_resid']]
-        mock_allvars.return_value = ret_val
-
-        expected_zs = [
-            [['y2'], ['y4']],
-            [['y2'], ['y3']],
-            [['y1'], ['y4']],
-            [['y1'], ['y3']]]
-        calc_zs = variable_permutations_for_iv_equations(self.factors, self.included_factors, self.transition_names,
-                                                         self.measurements, self.anchored_factors, 1, 1)[1]
-
-        assert_equal(calc_zs, expected_zs)
-
-
-class TestNumberOfIVParameters:
-    def setup(self):
-        self.factors = ['f1', 'f2', 'f3']
-        self.transition_names = ['bla', 'bla', 'blubb']
-        self.included_factors = []
-        self.measurements = []
-        self.anchored_factors = []
-        self.periods = []
-
-    @patch('skillmodels.estimation.wa_functions.variable_permutations_for_iv_equations')
-    @patch('skillmodels.estimation.wa_functions.tf')
-    def test_number_of_iv_parameters(self, mock_tf, mock_permut):
-        mock_tf.iv_formula_bla.return_value = ('1 + 2 + 3 + 4', '_')
-        ret = (['correct', 'wrong'], ['correct2', 'wrong2'])
-        mock_permut.return_value = ret
-
-        expected_param_nr = 4
-        calc_res = number_of_iv_parameters(self.factors, self.transition_names, self.included_factors,
-                                           self.measurements, self.anchored_factors, self.periods, 'f1')
-        assert_equal(calc_res, expected_param_nr)
-
-    @patch('skillmodels.estimation.wa_functions.variable_permutations_for_iv_equations')
-    @patch('skillmodels.estimation.wa_functions.tf')
-    def test_right_calls(self, mock_tf, mock_permut):
-        mock_tf.iv_formula_bla.return_value = ('1 + 2 + 3 + 4', '_')
-        ret = (['correct', 'wrong'], ['correct2', 'wrong2'])
-        mock_permut.return_value = ret
-        number_of_iv_parameters(self.factors, self.transition_names, self.included_factors, self.measurements,
-                                self.anchored_factors, self.periods, 'f1')
-        mock_tf.iv_formula_bla.assert_has_calls([call('correct', 'correct2')])
-
-
-class TestExtendedMeasCoeffs:
-    def setup(self):
-        self.factors = ['f1', 'f2']
-        self.transition_names = ['linear', 'constant']
-        self.measurements = {
-            'f1': [['y01', 'y02'], ['y11', 'y12']],
-            'f2': [['y03', 'y04'], []]}
-
-        coeffs = np.arange(0.6, 3.0, 0.2).reshape((6, 2))
-        cols = ['loadings', 'intercepts']
-        index_tuples = [(0, 'y01'), (0, 'y02'), (0, 'y03'), (0, 'y04'),
-                        (1, 'y11'), (1, 'y12')]
-        self.index = pd.MultiIndex.from_tuples(index_tuples)
-        self.storage_df = pd.DataFrame(coeffs, index=self.index, columns=cols)
-
-    def test_extended_meas_coeffs_no_constant_factor_and_intercepts_case(self):
-        coeff_type = 'intercepts'
-        calc_intercepts = extended_meas_coeffs(self.storage_df, self.transition_names, self.factors, self.measurements,
-                                               coeff_type, 0)
-        expected_intercepts = pd.Series(
-            data=[0.8, 1.2, 1.6, 2.0],
-            name='intercepts', index=['y01', 'y02', 'y03', 'y04'])
-        assert_series_equal(calc_intercepts, expected_intercepts)
-
-    def test_extendend_meas_coeffs_constant_factor_and_loadings_case(self):
-        coeff_type = 'loadings'
-        calc_loadings = extended_meas_coeffs(self.storage_df, self.transition_names, self.factors, self.measurements,
-                                             coeff_type, 1)
-        expected_loadings = pd.Series(
-            data=[2.2, 2.6, 1.4, 1.8],
-            name='loadings',
-            index=['y11', 'y12', 'y03_copied', 'y04_copied'])
-        assert_series_equal(calc_loadings, expected_loadings)
-
-
-class TestResidualMeasurements:
-    def setup(self):
-        intercepts = pd.Series(
-            [3.0, 2.0], name='intercepts', index=['m2', 'm1'])
-        loadings = pd.Series(
-            [2.0, 0.5], name='loadings', index=['m1', 'm2'])
-        self.side_effect = [loadings, intercepts]
-        d = pd.DataFrame(data=np.array([[5, 4], [3, 2]]), columns=['m1', 'm2'])
-
-        self.y_data = ['dummy', d, 'dummy']
-        self.storage_df = pd.DataFrame()
-        self.transition_names = []
-        self.factors = []
-        self.measurements = []
-
-    @patch('skillmodels.estimation.wa_functions.extended_meas_coeffs')
-    def test_residual_measurements(self, mock_extcoeffs):
-        mock_extcoeffs.side_effect = self.side_effect
-        expected_data = np.array([
-            [1.5, 2],
-            [0.5, -2]])
-        expected_resid = pd.DataFrame(
-            expected_data, columns=['m1_resid', 'm2_resid'])
-        calc_resid = residual_measurements(self.storage_df, self.transition_names, self.factors, self.measurements,
-                                           self.y_data, period=1)
-        assert_frame_equal(calc_resid, expected_resid)
-
-
-class TestWANorminfoDict:
-    def setup(self):
-        n = {}
-        n['f1'] = {'loadings': [['y1', 4], ['y2', 5], ['y3', 6]],
-                   'intercepts': [['y4', 7], ['y5', 8]]}
-
-        df = pd.DataFrame(data=[[None]] * 3, columns=['f1'])
-        self.identified_restrictions = {
-            'coeff_sum_value': df, 'trans_intercept_value': df}
-        self.normalizations = n
-        self.stagemap = [0, 1, 2, 2]
-
-    def test_wa_norminfo_dict(self):
-        expected = {'loading_norminfo': ['y2', 5],
-                    'intercept_norminfo': ['y5', 8]}
-        calculated = model_coeffs_from_iv_coeffs_args_dict(self.normalizations, self.stagemap,
-                                                           self.identified_restrictions, 1, 'f1')
-        assert_equal(calculated, expected)
 
 
 class TestBSMethods:
